chore: update project name, repository, license, and author details
Updated the project name from "superProductivity" to "task_timer_2" to reflect the new branding. Changed the repository URL to point to the new GitHub repository. Updated the license from MIT to MPL (Mozilla Public License) to align with new licensing requirements. Modified the author information to reflect the current maintainer.